requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2-binary
feedparser
Pillow
scrapy